Fixed: Now has winding correction for read in STL files
NOTE: This only works correctly when the file is supplied by with reasonable
face normals. These face normals need only be within 90 degrees out of phase
with the face winding to considered good enough to not require correction. That
is to say the face normal faces "outwards" within 90 degrees of the face's plane
and OpenGL logic agrees that is the front face.